move sites dropdown to node selector
