+++ lib/content-services/src/lib/content-node-selector/site-dropdown/sites-dropdown.component.ts
@@ -21,7 +21,7 @@ import { SitePaging, SiteEntry, Site } from '@alfresco/js-api';
 import { MatSelectChange, MatSelectModule } from '@angular/material/select';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
-import { SitesService } from '../common/services/sites.service';
+import { SitesService } from '../../common/services/sites.service';
 import { CommonModule } from '@angular/common';
 import { MatFormFieldModule } from '@angular/material/form-field';
