+++ lib/content-services/src/public-api.ts
@@ -21,7 +21,6 @@ export * from './lib/document-list/index';
 export * from './lib/content-user-info/index';
 export * from './lib/upload/index';
 export * from './lib/search/index';
-export * from './lib/site-dropdown/sites-dropdown.component';
 export * from './lib/breadcrumb/index';
 export * from './lib/version-manager/index';
 export * from './lib/content-node-selector/index';
